refactored trajectory code
