+++ wot/commands/__init__.py
@@ -10,3 +10,4 @@ from .force_layout import *
 from .transition_table import *
 from .trajectory import *
 from .util import *
+from .local_enrichment import *
