+++ wot/ot/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
-from .trajectory_sampler import *
 from .optimal_transport import *
-from .trajectory import *
-from .transport_map_summary import *
 from .optimal_transport_helper import *
+from .trajectory import *
+from .trajectory_trends import *
+from .transport_map_summary import *
 from .util import *
+
